+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Hitting `p` no longer triggers switching to the global search when focused on a link or a drop down menu
 - Fixed a bug that would disable navigation to the project page from the projects splash page on touch devices
 - Variables are now grouped correctly in the row and column dropdowns on the design report
+- Test subjects no longer have their status changed to valid if a sheet is entered for them and their subject code is in the site's valid range
 
 ## 0.12.3 (February 1, 2013)